MNT Fold ModalHeader back into Modal
