+++ client/src/components/Modal/Modal.js
@@ -1,8 +1,26 @@
 import React from 'react';
-import { Modal as ReactStrapModal } from 'reactstrap';
-
+import {
+  Modal as ReactStrapModal,
+  ModalHeader as ReactStrapModalHeader
+} from 'reactstrap';
+import i18n from 'i18n';
 import PropTypes from 'prop-types';
-import ModalHeader from './ModalHeader';
+
+/**
+ * The modal title can be pass as a string or an object with a html property.
+ * This utility method parse the title back to a string.
+ * @param string|object title
+ * @returns string
+ */
+function parseTitle(title) {
+  if (typeof title === 'object') {
+    // FormSchema title occasionally contains html, only render text for modal title
+    const doc = new DOMParser().parseFromString(title.html, 'text/html');
+    return doc.body.textContent || '';
+  }
+
+  return title;
+}
 
 /**
  * Component to render a modal
@@ -26,28 +44,51 @@ const Modal = ({
     modalClassName={modalClassName}
     size={size}
   >
-    <ModalHeader
-      onClosed={onClosed}
+    {title !== false && (
+    <ModalHeaderComponent
+      toggle={onClosed}
       title={title}
-      ModalHeaderComponent={ModalHeaderComponent}
-      showCloseButton={showCloseButton}
-    />
+    >
+      {parseTitle(title)}
+    </ModalHeaderComponent>
+    )}
+
+    {title === false &&
+    showCloseButton === true &&
+    typeof onClosed === 'function' && (
+      <button
+        type="button"
+        className="close modal__close-button"
+        onClick={onClosed}
+        aria-label={i18n._t('Admin.CLOSE', 'Close')}
+      />
+    )}
+
     {children}
   </ModalComponent>
 );
 
 Modal.propTypes = {
-  isOpen: PropTypes.bool,
   className: PropTypes.string,
+  isOpen: PropTypes.bool,
   modalClassName: PropTypes.string,
-  size: PropTypes.oneOf(['', 'sm', 'lg', 'xl']),
   ModalComponent: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
-  ...ModalHeader.propTypes
+  ModalHeaderComponent: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+  onClosed: PropTypes.func,
+  size: PropTypes.oneOf(['', 'sm', 'lg', 'xl']),
+  showCloseButton: PropTypes.bool,
+  title: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.bool,
+    PropTypes.shape({ html: PropTypes.string })
+  ]),
 };
 
 Modal.defaultProps = {
   isOpen: false,
   ModalComponent: ReactStrapModal,
+  ModalHeaderComponent: ReactStrapModalHeader,
+  title: null,
 };
 
 export default Modal;
